Refactor processing application into distinct components
