+++ bundles/processing/App.tsx
@@ -1,24 +1,60 @@
-import React from 'react';
+import * as React from 'react';
 import { Route, Switch } from 'react-router';
-import loadable from '@loadable/component';
+import { Accent, AppContainer, Theme } from '@spyrothon/sparx';
 
 import { useConstants } from '@common/Constants';
+import { usePermission } from '@public/api/helpers/auth';
 import { setAPIRoot } from '@public/apiv2/HTTPUtils';
+import { ProcessingSocket } from '@public/apiv2/sockets/ProcessingSocket';
 
-const ProcessingV2 = loadable(() => import('./ProcessingV2' /* webpackChunkName: 'processingV2' */));
+import { loadDonations } from './modules/donations/DonationsStore';
+import useProcessingStore from './modules/processing/ProcessingStore';
+import * as Theming from './modules/theming/Theming';
+import ProcessDonations from './pages/ProcessDonations';
+import ReadDonations from './pages/ReadDonations';
 
-function App() {
+import '../.design_system/generated/DesignSystem.css';
+import '@spyrothon/sparx/style.css';
+
+export default function App() {
+  const canChangeDonations = usePermission('tracker.change_donation');
+  const { processDonation } = useProcessingStore();
+  const { theme, accent } = Theming.useThemeStore();
+
   const { APIV2_ROOT } = useConstants();
 
   React.useEffect(() => {
     setAPIRoot(APIV2_ROOT);
   }, [APIV2_ROOT]);
 
+  React.useEffect(() => {
+    const unsubActions = ProcessingSocket.on('processing_action', event => {
+      loadDonations([event.donation]);
+      if (event.action !== 'unprocessed') {
+        processDonation(event.donation, event.action, false);
+      }
+    });
+
+    const unsubNewDonations = ProcessingSocket.on('donation_received', event => {
+      loadDonations([event.donation]);
+    });
+
+    return () => {
+      unsubActions();
+      unsubNewDonations();
+    };
+  }, [processDonation]);
+
   return (
-    <Switch>
-      <Route component={ProcessingV2} />
-    </Switch>
+    <AppContainer theme={theme as Theme} accent={accent as Accent}>
+      <Switch>
+        {canChangeDonations && (
+          <>
+            <Route path="/v2/:eventId/processing/donations" exact component={ProcessDonations} />
+            <Route path="/v2/:eventId/processing/read" exact component={ReadDonations} />
+          </>
+        )}
+      </Switch>
+    </AppContainer>
   );
 }
-
-export default App;
