--- bundles/processing/Processing.mod.css
@@ -1,72 +1,0 @@
-.container {
-  display: flex;
-  font-weight: initial;
-  background-color: var(--background-primary);
-  color: var(--text-primary);
-  height: 100vh;
-}
-
-.sidebar {
-  flex: 0 0 320px;
-  background-color: var(--background-secondary);
-  padding: 16px;
-  overflow-y: scroll;
-}
-
-.partitionSelector.partitionSelector > * {
-  /* Ensure the form inputs look balanced and don't squish the text */
-  flex: 1 1 min-content;
-}
-
-.main {
-  flex: 1 1 auto;
-  position: relative;
-  overflow-y: scroll;
-  padding: 0 16px;
-}
-
-.endOfList {
-  padding: 32px 0;
-  margin-bottom: 32px;
-  text-align: center;
-}
-
-.donationEnter {
-  opacity: 0;
-  transform: translateY(20px);
-  pointer-events: none;
-}
-.donationEnterActive {
-  opacity: 1;
-  transform: translateY(0);
-  transition: 120ms ease-out;
-  pointer-events: none;
-}
-.donationExit {
-  opacity: 1;
-  transition: 50ms ease-out;
-  pointer-events: none;
-}
-.donationExitActive {
-  transform: translateY(-30px);
-  opacity: 0;
-  pointer-events: none;
-}
-
-.groupsHeader {
-  display: flex;
-  align-items: center;
-}
-
-.groupsHeaderTitle {
-  margin-right: auto;
-}
-
-.modcommentTooltip {
-  white-space: pre;
-}
-
-.donationDragHandle {
-  flex: 0 0 auto;
-  cursor: grab;
-}
