+++ bundles/processing/modules/donations/HighlightKeywords.mod.css
@@ -1,0 +1,4 @@
+.highlighted {
+  background-color: #ff0;
+  color: #000;
+}
