+++ bundles/processing/modules/donations/ModCommentTooltip.mod.css
@@ -1,0 +1,3 @@
+.tooltip {
+  white-space: pre;
+}
