+++ bundles/processing/modules/donations/ModCommentTooltip.tsx
@@ -3,7 +3,7 @@ import { Clickable, Text, useTooltip } from '@spyrothon/sparx';
 
 import InfoCircle from '@uikit/icons/InfoCircle';
 
-import styles from './Processing.mod.css';
+import styles from './ModCommentTooltip.mod.css';
 
 interface ModCommentTooltipProps {
   comment: string;
@@ -12,7 +12,7 @@ interface ModCommentTooltipProps {
 export default function ModCommentTooltip(props: ModCommentTooltipProps) {
   const { comment } = props;
 
-  const [tooltipProps] = useTooltip<HTMLSpanElement>(<Text className={styles.modcommentTooltip}>{comment}</Text>, {
+  const [tooltipProps] = useTooltip<HTMLSpanElement>(<Text className={styles.tooltip}>{comment}</Text>, {
     attach: 'bottom',
     align: 'start',
   });
