+++ bundles/processing/modules/processing/ProcessingPartitionSettings.mod.css
@@ -1,0 +1,4 @@
+.partitionSelector.partitionSelector > * {
+  /* Ensure the form inputs look balanced and don't squish the text */
+  flex: 1 1 min-content;
+}
