+++ bundles/processing/modules/reading/ReadDonationsStores.tsx
@@ -1,0 +1,9 @@
+import create from 'zustand';
+
+interface ReadDonationsStoreState {
+  navigation: Array<[string, string[]]>;
+}
+
+const useReadDonationsStore = create<ReadDonationsStoreState>()(() => ({
+  navigation: [],
+}));
