+++ bundles/processing/pages/ReadDonations.mod.css
@@ -1,0 +1,8 @@
+.groupsHeader {
+  display: flex;
+  align-items: center;
+}
+
+.groupsHeaderTitle {
+  margin-right: auto;
+}
